+++ webonary-cloud-api/.vscode/settings.json
@@ -27,6 +27,7 @@
         "reversallanguages",
         "reversalletterheads",
         "semanticdomains",
+        "structs",
         "upsert",
         "upserted",
         "vscode",
