+++ webonary-cloud-api/lambda/getDictionary.ts
@@ -1,13 +1,8 @@
 import { APIGatewayEvent, Context, Callback } from 'aws-lambda';
 import { MongoClient } from 'mongodb';
 import { connectToDB } from './mongo';
-import {
-  DB_NAME,
-  DB_COLLECTION_DICTIONARIES,
-  DB_COLLECTION_ENTRIES,
-  PATH_TO_ENTRY_PART_OF_SPEECH_VALUE,
-  Dictionary,
-} from './db';
+import { DB_NAME, DB_COLLECTION_DICTIONARIES, DB_COLLECTION_ENTRIES } from './db';
+import { Dictionary, PATH_TO_ENTRY_PART_OF_SPEECH_VALUE } from './structs';
 import * as Response from './response';
 
 let dbClient: MongoClient;
