+++ webonary-cloud-api/lambda/postEntry.ts
@@ -215,16 +215,15 @@
 import { APIGatewayEvent, Callback, Context } from 'aws-lambda';
 import { MongoClient, ObjectId, UpdateWriteOpResult } from 'mongodb';
 import { connectToDB } from './mongo';
+import { DB_NAME, DB_COLLECTION_ENTRIES, DB_MAX_UPDATES_PER_CALL } from './db';
 import {
-  DB_NAME,
-  DB_COLLECTION_ENTRIES,
-  DB_MAX_UPDATES_PER_CALL,
   DictionaryEntryItem,
+  EntryAnalysisItem,
   EntryFileItem,
   EntrySenseItem,
   EntryValueItem,
-  copyObjectIgnoreKeyCase,
-} from './db';
+} from './structs';
+import { copyObjectIgnoreKeyCase } from './utils';
 import * as Response from './response';
 
 interface PostResult {
@@ -245,6 +244,7 @@ export async function handler(
   context.callbackWaitsForEmptyEventLoop = false;
 
   try {
+    const dictionaryId = event.pathParameters?.dictionaryId ?? '';
     const postedEntries = JSON.parse(event.body as string);
 
     let errorMessage = '';
@@ -254,7 +254,11 @@ export async function handler(
       errorMessage = `Input cannot be more than ${DB_MAX_UPDATES_PER_CALL} entries per API invocation`;
     } else if (postedEntries.find(entry => typeof entry !== 'object')) {
       errorMessage = 'Each dictionary entry must be a valid JSON object';
-    } else if (postedEntries.find(entry => !('guid' in entry && entry.guid))) {
+    } else if (
+      postedEntries.find(
+        entry => !('_id' in entry && entry._id) && !('guid' in entry && entry.guid),
+      )
+    ) {
       errorMessage = 'Each dictionary entry must have guid as a globally unique identifier';
     }
 
@@ -263,29 +267,43 @@ export async function handler(
     }
 
     const entryFileItemKeys = Object.keys(new EntryFileItem());
-    const entrySenseItemKeys = Object.keys(new EntrySenseItem());
     const entryValueItemKeys = Object.keys(new EntryValueItem());
     const updatedAt = new Date().toUTCString();
 
     const entries: DictionaryEntryItem[] = postedEntries.map((postedEntry: any) => {
-      const entry = new DictionaryEntryItem(postedEntry.guid as string, updatedAt);
+      const _id: string = postedEntry._id ?? postedEntry.guid;
+      const entry = new DictionaryEntryItem(_id, dictionaryId, updatedAt);
 
-      entry.dictionaryId = postedEntry.dictionaryId ?? postedEntry.dictionaryid;
       entry.letterHead = postedEntry.letterHead ?? postedEntry.letterhead;
 
       entry.mainHeadWord = copyObjectIgnoreKeyCase('mainHeadWord', entryValueItemKeys, postedEntry);
+
       entry.pronunciations = copyObjectIgnoreKeyCase(
         'pronunciations',
         entryValueItemKeys,
         postedEntry,
       );
+
+      entry.morphoSyntaxAnalysis = copyObjectIgnoreKeyCase(
+        'morphoSyntaxAnalysis',
+        Object.keys(new EntryAnalysisItem()),
+        postedEntry,
+      );
+
+      entry.senses = copyObjectIgnoreKeyCase(
+        'senses',
+        Object.keys(new EntrySenseItem()),
+        postedEntry,
+      );
+
       entry.reversalLetterHeads = copyObjectIgnoreKeyCase(
         'reversalLetterHeads',
         entryValueItemKeys,
         postedEntry,
       );
-      entry.senses = copyObjectIgnoreKeyCase('senses', entrySenseItemKeys, postedEntry);
+
       entry.audio = copyObjectIgnoreKeyCase('audio', entryFileItemKeys, postedEntry);
+
       entry.pictures = copyObjectIgnoreKeyCase('pictures', entryFileItemKeys, postedEntry);
 
       return entry;
