+++ webonary-cloud-api/lambda_api_doc/api_project.json
@@ -18,7 +18,7 @@
   "apidoc": "0.3.0",
   "generator": {
     "name": "apidoc",
-    "time": "2020-05-15T05:23:29.207Z",
+    "time": "2020-05-15T17:12:37.583Z",
     "url": "http://apidocjs.com",
     "version": "0.20.1"
   }
