+++ webonary-cloud-api/tools/post-legacy.ts
@@ -3,7 +3,7 @@
 import axios, { AxiosBasicCredentials, AxiosResponse, AxiosError, AxiosRequestConfig } from 'axios';
 import * as mime from 'mime-types';
 import * as fs from 'fs';
-import { Dictionary, DictionaryEntry, EntryFile } from '../lambda/db';
+import { Dictionary, DictionaryEntry, EntryFile } from '../lambda/structs';
 import fileGrabber from './fileGrabber';
 import { FlexXhtmlParser } from './flexXhtmlParser';
